[refactor]: updated data types for some config fields & added additional error handling

+++ ahrs-monitor/src/config/imu.rs
@@ -20,7 +20,7 @@ app_config! {
         /// Info about IMU metrics in IDTP payload.
         pub metrics: ImuMetrics,
         /// IMU device identifier.
-        pub device_id: u16,
+        pub device_id: u8,
         /// INDTP protocol mode.
         pub protocol_mode: u8,
     }
